simple (stupid) backmatching attempt; not a viable solution (known up front though) due to the fact that if a choicepoint at the start of the graph conflicts with a constraint at the end (in terms of searching), it'll backtrack the whole thing and start over.
back jumping is required for this; additionally, implementation isn't use cycle aware (doesn't know to try and force a negated
containment match).

Finally, design such as this is incapable of delaying resolution of a branch- virtuals being the kicker.
A decision on a virtual provider early on my be daft if later on one of the virtual providers is forced via hard constraints, thus
that stack must be put away, and revisited when only thing assumed branches remaining in the processing queue are virtual related atoms.

Exemption to this is 'real' virtuals, moreso metapkgs; this can be handled via a metapkg attribute on pkgs.