add sorter to itermatch, and optimize the piss out of everything else to use it. upshot of this, sorting occurs per repo.itermatch return- so for "prefer highest", we've got a set of iterables in sorted order. Enter util.iterables.iter_sort; func, given a sorting func, maintains iterable order, yielding the cumulative highest.
End result is the same as if you just used a sort on a list, but it pulls only as needed- since for any repo that uses metadata
based visibility filters, this means their *always* will be disk activity, thus converting matches to a list and sorting accesses
packages that may not even be needed.  Iteratively sort it in combination with iter_caching, and you access only what's needed.
For system testing, knocked sys time down by 66%.