mass renaming lib code from .bash to .lib
this makes the intention of the code clear (including that it must be
written as library, rather than script, code); finally this will be
used as a way to identify the sources and extract function names out
of it and generate a constant DONT_EXPORT_FUNCS list, rather than
trying to do it at runtime for the vast majority.

DONT_EXPORT_FUNCS will still have to be updated for dynamic code,
but that should be case by case.