fix make.defaults expansion to be PMS compliant
Rules of profiles have changed a fair bit since this code was first written,
now make.defaults has access to the parent environment and it's
required by PMS; as such, match the spec.

One difference to note here is that for incrementals, portage does
*not* expose the collapsed form of the incremental into each make.defaults;
this is a portage bug.  We adhere to the spec.

Finally, remove the ability to pass incrementals into OnDiskProfile; never
was useful, and doesn't fit w/ implementation required by PMS.