Finer grained masters: allow awareness of if it's intentionally empty
This is used to control later code that swaps in the default repo for
PMS repos, or for badly configured repos.

While this differs from portage in mask stacking, this is intentional.
No one should have an overlay while expecting gentoo masks to apply- if
they really want that, then they should set masters.