Convert to using chained exceptions
Via this, we can basically annotate what config was up to at the time-
a form of manual RTTI.  Importantly, this method preserves the raw
traceback, preserves exceptions of each level, *but* allows us
to ensure exceptions coming out of config are ConfigurationError
derived (or flat out bugs that should be dealt with, rather than
data errors).

Outputting still needs some better cleanup, but a fair bit more
information is now provided, and pconfig is enhanced a bit to
give better info for uncollapsable output.