switch away from eclass preloading to eclass caching
Downside, we parse a used eclass twice.  Upshot, we don't parse every
single eclass even if they aren't used- this is a literal shift for a
up to date cache from ~15s to ~5s, including ~10s reduction for an
empty cache.  Pretty much there are a lot of small eclasses either
not used, or very rarely used, thus this optimization.