teach the resolver to upgrade around weak blockers
Specifically, if it's weak, then try to merge a version that isn't
affected by the blocker.  This is enforced at the time the blocker
is encountered, although weak allows it to land whenever.

If the resolution succeeds, then pkgcore can work around the weak
blocker and ignore it- if it can't, then either the resolver is too
stupid, or there isn't an option, thus treat it as strong.