Fix conversion of options.
Specifically, deserialize the options into an array- do this
because we have to keep the indentation of the secondary column
at the same level (previously we weren't, and this broke the table).

Additionally, convert {arg1,arg2} into <arg1  | arg2> so that they're
properly handled by the rst machinery, and more importantly so we can
put spaces into it (helps for outputting).