Fix rollback support when reverting a replace affected by a blocker.
Specifically, if the state is blocker !a:1 w/ a replace operation of
a-1:1 -> a-2:2, that replace is fine.  Reverting the replace however
would fail w/ an assertion error due to !a:1 catching a-1:1 during
revert.

Thus track the state, and only through the error if the blocker state
for that package has somehow changed since we last looked (unlikely).