Update news, and trust env dumps we know just came from us.
For build phases of unpack through install, we know we /had/ to be
the one who generated it, and we know that it came from our exact
version of pkgcore; thus we can relax the scrubbing requirements,
instead doing a direct sourcing.

This basically knocks ~50% off small packages (bsdiff for example)
runtime, and 25% off diffball; effectively for local hardware testing,
a 2.5-3s semi-constant reduction.