Fix performance issues in __regex_filter_input.
Original code using egrep wasn't horrible in performance; the problem
was that the $PATH check was wrong and would fail if QA interceptors
were turned on.

In the process of debugging that, sed was found to be 33% faster than
egrep; a fallback to python was added (which is 1.5x slower than egrep);
that leaves the (previously active in this case) bash regex route that
was 4.5x slower than egrep.

Short version, --attr environment sped up by >6x.