conditionalize the output for the pkg version being replaced
Previously it would only be shown for pkgs being upgraded or downgraded.
This misses the case where a pkg switches between repos. Note that it
will only show the unchanged version not including the changed repo_id
if verbose output is disabled, but this signifies an underlying change
(more than just a straight pkg rebuild) and matches current portage
output.