add formatting support for showing '~', '*', and '#' pkg symbols
Similar to portage, the '~' symbol relates to a package with an unstable
keyword which will be shown unless ACCEPT_KEYWORDS contains the related
unstable keyword. Likewise, the '*' symbol is for pkgs with missing
keywords and the '#' symbol is for masked pkgs.

Note that pmerge currently only shows these symbols when verbose mode is
enabled while portage shows them all the time unless the '--quiet'
option is enabled.