fix(bash/exit-handling): don't pass args to ebuild death hooks
Currently only four eclasses (java-utils-2, multiprocessing, toolchain,
and chromium) and sys-apps/sandbox in the main tree appear to use
EBUILD_DEATH_HOOKS and none of them use the args that are passed in.

However, the multiprocessing death hook breaks when args are passed
since the hook calls `wait` and the args (in this case the specific
failure message when an ebuild using the multiprocessing eclass breaks)
gets appended to the command which only works with PID arguments.

Therefore, we'll just run the hooks without passing the generic failure
messages as args since they should be handled/shown by the die function
at the PM level anyway.