fix(bash/exit-handling): terminate non-helper subshells on failure
Previously subshells spawned by the build process such as those used via
the multiprocessing eclass weren't being terminated properly when die
was called. This potentially allows the build process to continue all
the way to to merging files onto the livefs even though part of the
build failed.

Now we throw SIGUSR1 for helper deaths (which enables a slightly
different handler/error message) and SIGTERM for all other subshells.

Note that this requires >=bash-4 for usage of $BASHPID but we're already
using it other places (such as for $PKGCORE_EBUILD_PROCESS_PID) without
any fallbacks so we don't care.