feat(pmerge): don't match virtual pkgs when matching against livefs
This matches portage behavior so when something like 'udev' is given as
an argument, with both 'sys-fs/udev' and 'virtual/udev' installed on the
system, pmerge selects 'sys-fs/udev' and ignores the virtual.

We assume virtuals are rarely the target when using individual pkg names
on the commandline. On the rare occasion they are explicitly being
installed, the full 'virtual/PN' can be used to specify virtuals with
pkg names colliding with other installed pkgs.