fix(ebuild/profiles): don't require config object parameter for OnDiskProfile
Requiring a config object to be passed breaks
BundledProfiles.create_profile from pkgcore.ebuild.repo_objs which
causes test failures in pkgcore-checks.

This currently assumes we don't need anything that uses the parent_paths
method from ProfileNode when working with portage-2 profiles otherwise a
ValueError will be raised since the repo map won't be created.

Clearly we need a better way to dynamically create the repo mapping when
it's first accessed within the ProfileNode class without having to play
with ConfigHint or work with the config object from other classes.