feature(eapi-bash): support generic command deprecation for EAPIs
This works similarly to the already existing banned command support.
Briefly explained, within the helpers directory a symlink is created
inside the related EAPI directory that points to the
internals/deprecated script and is named for the deprecated command.
This script then overrides the real command and when called during the
processing of an ebuild with a matching EAPI echoes a QA warning before
executing the intended command.