feature(eapi-bash): add a general method for deprecating or banning fcns
This allows us to specify function names in BANNED_FUNCS or
DEPRECATED_FUNCS variables in the related EAPI file to easily show QA
warnings for deprecated functions or error messages for banned
functions. The alternative would be explicitly defining function
overrides for every function in the related EAPI file which is verbose
and annoying.