doc: more readme tweaks
Mostly pointing to readthedocs instead of the dying/soon to be dead
pkgcore.org docs site. Note that we need to fix API and man page doc
generation for readthedocs, this will probably take reworking the build
a bit and perhaps adding a requirements.txt file so readthedocs knows
what dependencies are necessary for building docs.