try to fix conditional use dep default handling for pcheck
In particular, the current method for handling multilib style deps that
use both use dep defaults such as '(-)' as well as conditionalizing the
deps appears to cause pcheck to think the flags are enabled for the
related profile.

This is an issue since most of the related multilib flags are masked
(due to being for a different target arch then the current host) so when
pcheck tests if it can actually enable the requested flags (by seeing if
they intersect with the masked flags) it fails.

Really the use dep handling probably needs to be simplified or at least
documented better so external interfaces are cleaner and easier to
figure out.