add local use flag parsing support for pkgs
Adds a local_use attribute for pkg objects that is constructed when
parsing the metadata.xml file.

This was specifically implemented for pcheck so it can migrate away from
using a repo's profiles/use.local.desc file for local use flags and
instead use the flags defined per pkg in metadata.xml.