reorder ops for package replacements
Basically we should make sure to write out changes to the livefs before
doing pkg_postrm for the pkg being removed and pkg_postinst for the pkg
being installed.

This should make pkgcore work as expected when using "best_version" or
similar in the related pkg_* phases (see gentoo bug #535870).