triggers: fix info regen trigger when initially installing texinfo
On systems where texinfo isn't installed, the install-info binary
doesn't exist so the pre_merge phase wouldn't initialize the set of
mtimes properly causing the post_merge hook to fail since it would try
to iterate through NoneType.

Now we always store the mtime state before checking for the install-info
binary so the pre_merge phase should work as expected in the specific
case where texinfo is the pkg being initially installed on a system.