doc: add support for extracting long descriptions for man pages
We use the first line of the script module's docstring as the summary
and the remaining lines as the main description for the man page. This
allows us to have much longer descriptions that fluidly use the extra
info from the related module docstrings, once we add more docs. :)

Note that if the docstring only contains the summary the man page
description just uses that so it works for now when no extra docs exist
yet.