commandline: show available options on error for various argument types
This includes affects various user facing scripts such as pinspect and
pmerge that can take repos, pkgsets, formatters, etc as args for certain
use cases. Now if an unknown value is entered, the list of valid values
is returned along with the error.

Note this currently assumes the proper pluralization of the config type
is simply formed by tacking on an 's'. I don't think this assumption is
broken yet for current usage. Might have to return to it when that
assumption is wrong.