doc: first step towards having better man pages
Now we only generate each script's synopsis and options from the related
argparse code instead of generating the whole file. This allows us to
easily add content to the man pages.

Further down the road, we might move away from auto-generated man pages
for scripts entirely and instead rely on generating option and synopsis
dumps that can be compared manually against the current content of a
man page.  This is mostly because it would be nice to allow extra
content to be added for various subcommands which in the current
method is tricky without adding lots of extraneous content to the
code.