commandline: revert upstream argparse change for subparser namespaces
The upstream patch from issue #9351 (merged in python-2.7.9) seems to
have broken the control for allowing a namespace to be passed and used
by a subparser causing problems like issue #23058. We can remove this
once upstream figures out how to fix things.

For us, running with patch from #9351 merged in 2.7.9 means that
subparsers use totally different namespaces then their parents so
certain objects that we rely on being there, e.g.  "config", are missing
since they're only in the parent namespace. This cases parsing to fail
when delayed objects relating to subparsers try to extract their config
info only to find it missing.

http://bugs.python.org/issue9351
http://bugs.python.org/issue23058