plugin: restrict regeneration to related module location only
Previously running pkgcore from a git checkout caused the system location for
pkgcore plugins to be regenerated as well if pkgcore was installed as a system
lib.

This fails if we're running pkgcore as a regular user since they don't have
write access to the system location. In addition, we probably don't want to
regen the system location when using a git checkout since we might rev the
cache version at some point so the system version might not be compatible with
the git version until the next release.

Fixes #35.