profiles: fix portage-2 profile support when using pinspect
This is still a giant hack, but at least it's now a giant hack that
should work for pinspect as well (and fixes #5 as such).

To improve the situation we should come up with a way to dynamically
create the repo_id -> repo_config mapping when the related attr is
accessed (especially since the portage-2 profile format isn't used by
default when running the standard gentoo repo) and then cache it for
further accesses. Off the top of my head, we could probably do something
using the snakeoil.klass optimizations along with a custom class
leveraging __call__. We might be able to work it into the current config
framework so it collapses and reuses existing config objects instead of
creating new objects.