add support for set-based profile-formats settings
Apparently this is how portage was always handling the setting in that
you can have more than one type set. Generally appears to be used to
extend support for various features outside of the spec.

We need it for supporting profile-formats options such as
profile-bashrcs which is generally set in combination with other formats
such as portage-2.