commandline: handle SIGINT better with spawned processes
Sending a single SIGINT signal should kill the whole process group now
while still allowing the relevant backtrace to be printed if '--debug'
is passed as an option.

If a spawned process also spawns python processes or others where SIGINT
is modified, then multiple SIGINTs must be sent in order to terminate
the entire process group under the usual scheme.

An example of this would be building pkgcore using emerge/pmerge and
hitting Ctrl-C when the 2to3 conversion process is running.

Now we reset SIGINT to the default and send it to the main process group
ID to properly interrupt all child processes.