Fix hardlinking induced bug in splitdebug resulting in $IMAGE content.
The change is fairly obvious in reading it; the original code was
incorrectly setting the fsobj location to the data source path- rather
than changing the location (which is what we primarily use).

This has been causing issues for a while most likely; anytime a hardlink
was encountered by splitdebug this would occur.  This was exposed due to
increased hardlink support- effectively we preserve hardlinks well enough
now that it was flushed out for binpkgs.