drop warnings if wheel and/or the portage users/groups are missing
These types of messages spew forth annoyingly on non-Gentoo systems
where pkgcore or pkgcheck work fine running as root or another user
(e.g. on travis-ci) or in a prefix setting once support is finished for
it.

Really it seems like it would better to document such requirements for
certain use cases and error out if absolutely necessary in the code
somewhere else.