commandline: add --override-config option
This makes it easier to run all the command line tools under different
configurations, e.g. maintaining separate configurations to target
different binpkg repos.

For portage compat, the path to the custom config directory should be
passed, otherwise if the path to a file is used it will try to load it
as a pkgcore config file.