don't assume bash is at /bin/bash, inspect path to find it
This allows us to more easily override the bash version used with a
custom, patched version if trying to debug something. Also, if we want
to support other system types like BSDs in the future its useful as well
since its likely to be found at /usr/local/bin there.