bash: clarify why expand_aliases still needs to be enabled
In summary, it's due to still needing aliases to properly load
portage-style env dumps that use 'declare' formatting which is
inherently local in scope (except if -g is given).

Loading them normally in a function would localize all the variables to
that function so nothing else could use them outside that function's
scope.