fix SIGINT handling due to process group ID changes
Previously the main python process was the process group leader, but
that was changed to make each spawned ebuild processor its own process
group leader so related child processes could be terminated on die()
properly.

Therefore, now we iterate through all the active and inactive ebuild
processors sending SIGINT to them so the signal percolates down to their
children in the hope that everything is interrupted as expected in a
quick and orderly fashion.