bash: fix dying in helpers
Previously both SIGUSR1 would be sent to the ebuild process and SIGTERM
to the ebd process group (PPID) which doesn't exist in the helper's
context (as well as __ebd_write_line()) so error messages alluding to
that fact would be output while dying.

This might break instances where helpers are subshelled with die calls
but we need better test suites in order to check for more cases like
that.