simplify the wrapper script functionality
This rids us of the pwrapper/pwrapper_installed linking method. Now the
initialization code lives in the main pkgcore.scripts module which
handles sys.path munging for repos and tarballs as well as being used as
a target for installed scripts that are generated during the build
process.

We could trivially use entry_points from setuptools to generate and
install the scripts as well, but that appears to add a relatively
significant amount of extra runtime for scripts executing under a second
due to all the extra modules pkg_resources imports. Also, we probably
don't need the extra complexity and layers of abstraction anyway.