bash: fix EXPORT_FUNCTIONS when passed multiple words as one string arg
For example, the current selinux-policy-2.eclass has the following
setting:

EXPORT_FUNCTIONS "src_unpack src_prepare src_compile src_install pkg_postinst pkg_postrm"

which in the previous implementation would try to make a single function
named as that entire string. Now we split each function name into a
separate arg whether or not the entire argument or pieces of it are
quoted. This should work because functions can't have spaces in them.