docs: initial work towards supporting extended docs for cli tools
In short, the 'docs' keyword arg can be used in add_argument() calls
from which the text will be placed under the related option or argument
in the man page.

We also hijack argparse's add_argument() function to make sure it
doesn't complain about the unsupported keyword arg and to avoid the
extended docs from showing up in the help output summary when running
tools with the -h/--help options.