pmerge: treat all package restrictions the same when parsing targets
Previously boolean 'and' restrictions (restriction combinations, e.g.
'>gnupg-1' would be parsed into a package dep and a version dep combined
together with an 'and' dep) wouldn't be treated the same even when they
included sub-restrictions that were package deps.

This caused unconsistent behavior in the case were there were multiple
matches, but only once was installed. Specifically if just a package
name was specified the installed pkg would be returned; however, if a
package name and version were specified then both packages would be
selected for merging.

Now the installed package should always be the only choice returned in
all cases where the restriction contains a package dep.