repository: use realpath instead of abspath to determine file repo status
Previously using abspath would cause a mismatch in certain cases for
files located in a symlinked repo since abspath actually resolves
symlinks in cases due to os.getcwd() also resolving symlinks when within
a path containing them.

Now when doing path comparisons to determine if a file is in a repo or
not we use both the file and the repo's real paths so all symlinks
should be resolved.