pmaint: allow more argument types for pmaint digest
Now you can run it without arguments within a directory in the repo and
it should remanifest everything that's either in the CWD or lower in the
hierarchy. You can also point it at an ebuild file to do single ebuild
manifests.

This also includes some small simplifications for the repo digesting
backend support, with much more still needed.