fetch: raise FetchFailed if distfile is empty
This fixes the case where the first spawn of wget returns an error (such
as a 404) and wget still creates an empty file in the distdir.

Previously we would assume everything is fine as long as the file
existed, but now we force the file to actually contain some kind of
content since we're relying on the verify process to tell us if fetching
passed/failed and currently not checking the return status of the
spawned fetch command.

This fixes `pmaint digest` not working properly when fetching distfiles
where the first URI (e.g. the 'to-be-populated' gentoo mirror) wouldn't
exist yet.