util/commandline: add workaround for py3 argparse required args changes
Subparsers default to required in py2 and optional in py3 (see
http://bugs.python.org/issue9253).

In general, when we use them one must always be specified so force them
to be required.

Fixes running `pmaint` or similar by itself producing unhandled
exceptions instead of an error message that a subcommand arg is
required on py3.