rework external eapi object support
Trying to simplify and make the API more consistent.

In summary, drop the eapi_obj attributes and rely on just eapi
attributes. Also, mark the magic and parent attrs of the EAPI instances
as private. If an external string representation of an EAPI object is
needed just do str(eapi) or similar as the EAPI class has __str__
defined.