ebuild.misc: only rebuild globals for true globals
Currently some package restrictions are making it into this method call
such as category globbed package.use settings which was causing
erroneous use renderings.

The underlying issues in update_from_stream() where the conditional
calls add_global() if the item isn't an atom or similar should probably
be fixed instead in the future.