ebuild-daemon: Fix command flow conditionality on SANDBOX_LOG
Avoid having the ebuild-daemon startup command flow conditional on
SANDBOX_LOG variable. If pkgcore is run inside external sandbox,
the SANDBOX_LOG variable will be set even though pkgcore did not enable
sandbox. As a result, ebuild-daemon will hang waiting for the command,
while processor will hang waiting on the output following the block.

To solve that issue, reflow the process to have processor always send
a command -- indicating whether the sandbox is enabled or not.
In the former case, the ebuild-daemon replies with the log file as
before. In the latter case, it just continues execution ignoring
SANDBOX_LOG.