pmerge: parse_target(): remove virtual pkgs if no installed matches are found
This helps match emerge behavior from portage when multiple pkgs have
the same name including a virtual, but none are installed. In this case,
the virtual pkg should be ignored and the regular pkg should be prepped
for install.

Previously, pmerge would complain that multiple pkgs were found in that
case.