pmerge: add initial -d/--downgrade support
Resolves with intent opposite to -u/--upgrade.

Useful for reverting to the most recent binpkg not matching the current
installed version or rebuilding the previous ebuild if no binpkgs exist.

In other words, it's shorthand for having to enter in specific atom
versions when wanting a quick revert to a previous version and as such
it's most useful when used with -O/--nodeps otherwise lots of resolved
downgrades get pulled in as well.