Fix pmaint sync buffering and flushing.
The code was written assuming that the out/err is unbuffered-
which is not true.  That in turn leads to the underlying repo sync operations
(which are usually raw processes writing to the fd) having output intermixed.

Thus just force a flush before handing off to the repo operations, and sync upon
return from it.

Additionally, while I'm in here, cleanup the output a bit and drop the
fetish for forcing output to stderr.