Add comment explaining why https is used for sqfs
http is a joke in 2020, but distfiles.gentoo.org doesn't use
either redirects to upstreams, or valid SAN for certs, so while it
speaks tls- the PKI is invalid, rendering this a bit pointless.

Add a comment to clarify for anyone else who comes across it as I did,
explaining that it's an intentional choice (despite it being daft).