Update http syncer to more reliably handle 304.
If a 304 is returned, this is actually an exception- which the code
was just bleeding through to the invoker as an exception instead of
handling.

Additionally, the type conversions for etag/modifier comparison for a
200 return- which frankly I'm not sure exists (it's a broken web server)-
didn't have aligned type conversions, or basics like space suppression.  While
I'm in here, I'm adding that, and replacing some boilerplate with readfile
usage.