Refactor _get_attr logic into a more readable metaclass.
Per the docstring of DynamicGetattrSetter, there still is a performance
gain to the original "abuse __getattr__ as a hook to generate the attr,
and then leverage the cpy/vm to fastpath subsequent accesses"; it still
is faster in comparison to `snakeoil.klass.jit_attr`, and at least one
check in pkgcheck hardcoded awareness of _get_attr, so it's not trivial
to remove.

What this metaclass allows is to rewrite the package class definitions
into a form that basically reads like `@property` or `@klass.jit_attr`
style invocations, and the DynamicGetattrSetter metaclass mangles things
in the background to maintain compatibility.

Said another way; the code is more readable, same performance, and it
refactors things so this machinery can be replaced with `jit_attr` if
the c extension of jit_attr is updated (or one decides they don't care
about the relatively small speedup this still provides).