pquery: revert profile override setting
Pquery currently defaults to querying configured/filtered repos. This
profile override breaks that assumption since most of the regular
profile settings aren't available in it.

It might be nice to reverse that in the future by making it query
against raw repos by default and optionally apply configuration
filtering, but we'll stick with the status quo for now.