ebuild.repository: rewrite pkg manifesting support
Defer to pkgcheck to catch errors previously checked for during
manifesting and only manifest the fetchables for the ebuilds matching
the specified restriction instead of all packages matching the
unversioned atoms for the restriction. This allows selective manifest
updates via `pkgdev commit` and `pkgdev manifest` instead of using all
the ebuilds for a given package.