ebd_ipc: Fix unpack functionality broken via refactoring.
Rough tracing takes this back to logic being added to allow
for fetchers that have differing distdirs; the problem is
there is a break in the encapsulation.

Basically the fetcher has to return the location of the
content on disk, but just returns a boolean indicating success.

Being the invoking operation (setup usually) wasn't getting
updated for where the distfiles where (becaused the `verified_files`
attribute wasn't updated), unpack functionality was broken.

This represented via files being downloaded, but not being symlinked
into the spoof DISTDIR; thus no unpack.