ebuild/portageq.py: deduplicate get_repo_news_path
On Python 3.11, pquery run fails as we have subparser conflict -
`get_repo_news_path` is defined 2 times under the same name. The first
one is the original while the second one has "repo" as optional.
In the implementation we can see that the second one just calls the
first one, meaning if repo wasn't passed, it will fail!

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>