fetch.custom: fix commands with `\` variable escaping
When passing fetch commands, such as `FETCHCOMMAND` or `RESUMECOMMAND`,
a user might want to use `\` escaping, such as in the resolved issue.
pkgcore was mistakenly replacing only the var name, while leaving the
`\` in place, which would cause the command to fail. Now this matches a
little more portage's behavior.

Resolves: https://github.com/pkgcore/pkgdev/issues/94
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>