ebuild.repository: improve handling with precise restricts
When the restricts used for manifest return subset of versions of the
package, calculating the manifest drops all entries for other unmatched
versions.

We separate the behavior into 2 types: selected and unselected. For the
selected once by restricts, the behaviour shouldn't change at all - if
it did, this is a bug! For the unselected once, it should reuse the
current manifest entries. In case a distfile has missing checksums of
the required types, it will drop the entry! This behavior should occur
only if the current Manifest is broken, so we allow ourselves to do what
we want.

Resolves: https://github.com/pkgcore/pkgdev/issues/74
Resolves: https://github.com/pkgcore/pkgdev/issues/83
Resolves: https://github.com/pkgcore/pkgdev/issues/101
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>