pmerge: sort target processing to stabilize the graph.
The current resolver is effectively brute force graph exploration,
iteratively adding results to the graph changing previously decisions
based on new requests.  This means that the resolver *can* produce
different solutions if the order of constraints added to it differ.

It's desirable the resolver behaviour be reproducible, thus change the default
behaviour to sort the target list given to the resolver.

As to why we don't force sorting within the resolver itself; for processing
of package dependencies, those are already stably ordered.  A future
enhancement would be to modify the resolver to be explicit, but that's for
when the resolver is rewritten.

Currently the source of instability is the atoms fed to the resolver from
pmerge, thus why I'm adding the sorting in pmerge.

Finally: the resolver must always produce a solution, or must always not, irregardless
of the argument ordering it's given.  Thus '--disable-resolver-target-sorting' is
added to allow disabling the sorting and doing something of a random entry/walk
into the depgraph.  This can produce differing solutions than what a sorted entry
would produce; that is working as intended (there is no global optimizer of solutions,
thus 'first solution' is what pmerge uses).

Basically, these two invocations should agree that a solution exists, or doesn't.
If they disagree this directly shows that the resolver is failing to explore solution
space fully:
* `pmerge <targets> --disable-resolver-target-sorting`
* `pmerge <targets>

Signed-off-by: Brian Harring <ferringb@gmail.com>