Fix secondary exception from when an EBD fails to start.
Further in this __init__, self.pid is initialized to the process
that is started.  If that fails- an exception is thrown- then
self.pid doesn't exist, which in turn breaks the assumptions of
`shutdown_processor()` and `.is_alive` (both of which are
designed to work if an init partially failed).

Finally, add protections to ensure a failed spawn doesn't leak
file descriptors.  Python won't reap raw FD's, so we need to
handle this gracefully.

This secondary exception is visible in pkgcore/pkgcore#296