Add a constant of the latest PMS version, and add use flag validation to EAPI.
Currently no EAPI varies the use flag parsing rules, but if it ever comes up,
this will address it.

In the process, this also will be used to move some logic out of atom.py whilst
introducing the concept of 'latest PMS' as the 'latest' rules to parse under in
the absense of an explicit directive.

Signed-off-by: Brian Harring <ferringb@gmail.com>