Move valid_use_flag parsing out of atom, rely upon eapi obj.
EAPI objects are responsible for deciding the rules of how to parse
atoms, and this extends to use flags.  Thus just ask the EAPI obj
to tell us "is it valid?".

In moving this out of atom.py it allows us to lace that validation
into other configuration loading aspects of the code.

Signed-off-by: Brian Harring <ferringb@gmail.com>