fix(sync): Add loggger.debug() of the sync command actually being invoked.
Additionally, since we know the sync subsystem was written before basic sanity
of the frontend, and we know that it invokes processes that write to stdout/stderr,
force a flush of those handles to ensure that we don't wind up with interlaced output
from two processes.

The need to bury a flush here is a sign that the abstraction needs redesign, but
that's a problem for a later date.

Signed-off-by: Brian Harring <ferringb@gmail.com>
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>