refactor(portage_conf): Push fallback (and path awareness) logic of repos.conf into methods.
The previous code had the exception + fallback handling in the init; this is already
a long init, and there are changes needed for repos conf, thus introduce a method to
do the fallback logic.

Said method will be extended to essentially bury the complexity of repos.conf- that problem-
into an encapsulated method to try and reduce the cognitive burden of dealing with it.

Why this matters: this work is prep for extension of repos.conf to add in various syncing
features portage grew over the last decade+.

Signed-off-by: Brian Harring <ferringb@gmail.com>
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>