refactor(config): adding typing to config hints along with immutability.
ConfigHint's should be treated as immutable; any manipulation of their data
should be handled either via clone or via pre-processing.  This system is
complex enough we don't want mutability, and the allowance of that was an
oversight from the good ole days of py2.4.

Force immutability sincei t's the actual design contract and also to flush
out anyone breaking said contract (no pkgcore code breaks that, to be clear).

Signed-off-by: Brian Harring <ferringb@gmail.com>