refactor(config): remove the notion of config layer incremental prepend/append.
Many moons gone past the original config system tried to support the idea
of incrementally stacked values (x y -x z -y == z).  This commit removes
the fundamental core of that functionality; the incremental logic is now implemented
in domain specific layers, rather than being a capability of the config layer.

Note: this just removes the logic.  The signature of [before, mid, after] return
is still in place, and will be removed as the API is normalized.  In particular,
pconfig still reports on .prepend/.apend, there just no longer is content there.

That behaviour (and signature) change will be a followon commit.

Signed-off-by: Brian Harring <ferringb@gmail.com>