refactor(config): remove prepend/append incremental return signature fully.
The previous commit disabled the logic to generate prepend/val/append- nothing used it-
this commit realigns the client consumers to expect a single value unless they're invoked
arg_type="repr".

Essentially, now if you ask the config layer for a value, you don't have to know if it's
a value that is considered 'incremental'- you just get back a python object for what you asked,
ready for usage.

Signed-off-by: Brian Harring <ferringb@gmail.com>
Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>